backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.14
django-filter==24.2
django_extra_fields==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docker==7.0.0
flake8==7.1.0
gunicorn==22.0.0
idna==3.7
npm==0.1.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
Pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
python3-openid==3.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19